requirements.txt
python-dotenv
openai
PyPDF2
google-generativeai
Flask